html/browsers/browsing-the-web/navigating-across-documents/007.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<link rel="help" href="https://html.spec.whatwg.org/multipage/links.html#following-hyperlinks">
<title>Link with onclick that sets location to "javascript: url" is aborted by href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="document.querySelector('#test').contentWindow.location = 'javascript:`<script>self.postMessage(&quot;click&quot;, &quot;*&quot;);</script>`';" href="href.html">Test</a>
<script>
var test = async_test();

window.onmessage = test.step_func_done(function(e) {
  // Navigating the iframe to the anchor's `href` will abort the previous operation.
  // If the semantics are implemented correctly, a message "href"
  // from "href.html" will be received.
  assert_equals(e.data, "href");
});

test.step(function() {
  // Clicking the anchor will trigger the `onclick` handler operation,
  // which will attempt to set the `location` of the iframe to a "javascript: url",
  // that creates a script element containing a `postMessage` call, which will
  // send a message "click".
  document.querySelector("a").click();
});
</script>
